Can close modal with esc key or by clicking outside of modal
