+++ package.json
@@ -36,6 +36,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-icons": "^4.2.0",
+    "react-spring": "^9.2.4",
     "react-test-renderer": "^17.0.2",
     "styled-components": "^5.3.1",
     "webpack": "^5.52.1"
